[Sdk] Separated nested ProfilesLibrary classes and added IsLoaded method (#175)
