+++ FrostySdk/ProfileVersion.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FrostySdk
+{
+    public enum ProfileVersion
+    {
+        NeedForSpeedRivals = 20131115,
+        Battlefield4 = 20141117,
+        DragonAgeInquisition = 20141118,
+        NeedForSpeed = 20151103,
+        StarWarsBattlefront = 20151117,
+        PlantsVsZombiesGardenWarfare = 20140225,
+        PlantsVsZombiesGardenWarfare2 = 20150223,
+        MirrorsEdgeCatalyst = 20160607,
+        Fifa17 = 20160927,
+        Battlefield1 = 20161021,
+        MassEffectAndromeda = 20170321,
+        Fifa18 = 20170929,
+        NeedForSpeedPayback = 20171110,
+        StarWarsBattlefrontII = 20171117,
+        Madden19 = 20180807,
+        Fifa19 = 20180914,
+        Battlefield5 = 20180628,
+        NeedForSpeedEdge = 20171210,
+        Anthem = 20181207,
+        Madden20 = 20190729,
+        PlantsVsZombiesBattleforNeighborville = 20190905,
+        Fifa20 = 20190911,
+        NeedForSpeedHeat = 20191101,
+        StarWarsSquadrons = 20201001
+    }
+}
